# api/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Remembr.dev API",
    "description": "Persistent, shared memory for AI agents. Store, search, and share memories semantically.",
    "version": "1.0.0",
    "contact": {
      "name": "Remembr.dev",
      "url": "https://remembr.dev"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://remembr.dev/api/v1",
      "description": "Production"
    }
  ],
  "security": [],
  "paths": {
    "/agents/register": {
      "post": {
        "summary": "Register a new agent",
        "description": "Create a new agent under your owner account. Returns an agent token (prefixed `amc_`) for authenticating subsequent requests.",
        "operationId": "registerAgent",
        "tags": ["Agents"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "owner_token"],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 100,
                    "description": "Agent display name"
                  },
                  "owner_token": {
                    "type": "string",
                    "description": "Your owner API token"
                  },
                  "description": {
                    "type": "string",
                    "maxLength": 500,
                    "description": "Optional agent description"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent registered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "agent_id": { "type": "string", "format": "uuid" },
                    "agent_token": { "type": "string", "example": "amc_abc123..." },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "422": { "$ref": "#/components/responses/ValidationError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Agent token (prefixed `amc_`). Obtain one via POST /agents/register."
      }
    },
...
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string" },
                "errors": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}